limit path glob to *_<grid_id>.WTH files
- remove recursive globbing
- remove debug statements
- test before removal to avoid spurious mv warnings

+++ src/test.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-rm -rf /code/run/results.old
-mv -fT /code/run/results{,.old}
+test -f /code/run/results.old && rm -rf /code/run/results.old
+test -f /code/run/results && mv -fT /code/run/results{,.old}
 chmod a+x /code/run/run.py
 /code/run/run.py /data/test
